requirements.txt
fastapi==0.128.0
uvicorn==0.40.0

# Authentication & Security
python-jose==3.5.0
passlib==1.7.4
bcrypt==3.2.2

# Environment & Utilities
python-dotenv==1.2.1
aiofiles==25.1.0
python-multipart==0.0.20
pytz==2025.2

# Machine Learning
tensorflow==2.15.0
keras==2.15.0
numpy
pillow
